Fixed issue with non existing files

+++ index.js
@@ -47,7 +47,8 @@ function saveDnsEntry(data) {
 }
 
 function clearExistingFile(filePath){
-  fs.unlinkSync(filePath);
+  if(fs.existsSync(filePath))
+    fs.unlinkSync(filePath);
 }
 
 function getFilename(){
